User roles added

+++ frontend/script.js
@@ -1,5 +1,6 @@
 const jwt = localStorage.getItem("token");
 const username = localStorage.getItem("username");
+const role = localStorage.getItem("role");
 const container = document.getElementById("books-wrapper");
 let books = [];
 
@@ -30,12 +31,19 @@ const deleteBook = async (id) => {
 const main = async () => {
   await getBooks();
 
-  books.forEach((book) => {
-    let newBook = `<div class="swiper-slide box book" style="margin-left: 10px;
+  if (role == "ADMIN") {
+    const addBookDiv = document.getElementById("addBookDiv");
+    addBookDiv.innerHTML = `<a class="pointer-cursor" href="./addBook/addBook.html">
+        Add book
+      </a>`;
+
+    books.forEach((book) => {
+      let newBook = `<div class="swiper-slide box book" style="margin-left: 10px;
     margin-right: 10px;">
     <div class="icons">
       <a onClick="deleteBook(${book.id})" class="fas fa-trash-can"></a>
       <a href="./updateBook/update.html?id=${book.id}" class="fas fa-pen-to-square"></a>
+      <a class="fas fa-cart-plus"> </a>
     </div>
     <div class="image">
       <img src="http://localhost:4004/uploads/${book.coverFile}" alt="" />
@@ -46,8 +54,27 @@ const main = async () => {
       <p>Genre: ${book.genre}</p>
     </div>
   </div>`;
-    container.innerHTML += newBook;
-  });
+      container.innerHTML += newBook;
+    });
+  } else {
+    books.forEach((book) => {
+      let newBook = `<div class="swiper-slide box book" style="margin-left: 10px;
+    margin-right: 10px;">
+    <div class="icons">
+      <a class="fas fa-cart-plus"> </a>
+    </div>
+    <div class="image">
+      <img src="http://localhost:4004/uploads/${book.coverFile}" alt="" />
+    </div> 
+    <div class="content">
+      <h3>${book.title}</h3>
+      <p>${book.description}</p>
+      <p>Genre: ${book.genre}</p>
+    </div>
+  </div>`;
+      container.innerHTML += newBook;
+    });
+  }
 };
 
 const usernameField = document.getElementById("username-field");
